#19 > folder 구조 변경 >  account controller를 페이지별 controller로 변경
