+++ src/main/java/com/kh/workPeople/account/login/controller/LoginController.java
@@ -1,0 +1,25 @@
+package com.kh.workPeople.account.login.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("/account")
+public class LoginController {
+
+	@GetMapping("/login")
+	public String login() {
+		return "account/login/login";
+	}
+	
+	@GetMapping("/manager-login")
+	public String mangerLogin() {
+		return "account/login/login-manager";
+	}
+	
+	@GetMapping("/changePassword")
+	public String changePassword() {
+		return "account/join/change-password";
+	}
+}
